// Day11/src/types.ts
export interface IGithubEvent {
    id:         string;
    type:       string;
    actor:      Actor;
    repo:       IGithubEventRepo;
    payload:    Payload;
    public:     boolean;
    created_at: Date;
    org?:       Actor;
}

export interface Actor {
    id:             number;
    login:          DisplayLoginEnum;
    display_login?: DisplayLoginEnum;
    gravatar_id:    string;
    url:            string;
    avatar_url:     string;
}

export enum DisplayLoginEnum {
    Alex = "alex",
    Fishinabarrel = "fishinabarrel",
    Pyca = "pyca",
}

export interface Payload {
    ref?:           string;
    ref_type?:      string;
    pusher_type?:   string;
    master_branch?: Branch;
    description?:   string;
    action?:        Action;
    number?:        number;
    pull_request?:  PayloadPullRequest;
    push_id?:       number;
    size?:          number;
    distinct_size?: number;
    head?:          string;
    before?:        string;
    commits?:       Commit[];
    review?:        Review;
    issue?:         Issue;
    comment?:       Comment;
}

export enum Action {
    Closed = "closed",
    Created = "created",
    Opened = "opened",
}

export interface Comment {
    url:                      string;
    html_url:                 string;
    issue_url:                string;
    id:                       number;
    node_id:                  string;
    user:                     User;
    created_at:               Date;
    updated_at:               Date;
    author_association:       AuthorAssociation;
    body:                     string;
    reactions:                Reactions;
    performed_via_github_app: null;
}
// ...
export interface Commit {
    sha:      string;
    author:   Author;
    message:  string;
    distinct: boolean;
    url:      string;
}

export interface Author {
    email: string;
    name:  NameEnum;
}

export interface Issue {
    url:                      string;
    repository_url:           string;
    labels_url:               string;
    comments_url:             string;
    events_url:               string;
    html_url:                 string;
    id:                       number;
    node_id:                  string;
    number:                   number;
    title:                    string;
    user:                     User;
    labels:                   Label[];
    state:                    State;
    locked:                   boolean;
    assignee:                 null;
    assignees:                any[];
    milestone:                null;
    comments:                 number;
    created_at:               Date;
    updated_at:               Date;
    closed_at:                null;
    author_association:       string;
    active_lock_reason:       null;
    body:                     string;
    reactions:                Reactions;
    timeline_url:             string;
    performed_via_github_app: null;
    draft?:                   boolean;
    pull_request?:            IssuePullRequest;
}
// ...
export enum Branch {
    Main = "main",
    Master = "master",
}

export interface PayloadPullRequest {
    url:                    string;
    id:                     number;
    node_id:                string;
    html_url:               string;
    diff_url:               string;
    patch_url:              string;
    issue_url:              string;
    number:                 number;
    state:                  State;
    locked:                 boolean;
    title:                  string;
    user:                   User;
    body:                   null | string;
    created_at:             Date;
    updated_at:             Date;
    closed_at:              Date | null;
    merged_at:              Date | null;
    merge_commit_sha:       null | string;
    assignee:               null;
    assignees:              any[];
    requested_reviewers:    any[];
    requested_teams:        any[];
    labels:                 Label[];
    milestone:              null;
    draft:                  boolean;
    commits_url:            string;
    review_comments_url:    string;
    review_comment_url:     string;
    comments_url:           string;
    statuses_url:           string;
    head:                   Base;
    base:                   Base;
    _links:                 PullRequestLinks;
    author_association:     AuthorAssociation;
    auto_merge:             AutoMerge | null;
    active_lock_reason:     null;
    merged?:                boolean;
    mergeable?:             null;
    rebaseable?:            null;
    mergeable_state?:       string;
    merged_by?:             User | null;
    comments?:              number;
    review_comments?:       number;
    maintainer_can_modify?: boolean;
    commits?:               number;
    additions?:             number;
    deletions?:             number;
    changed_files?:         number;
}
// ...
export interface Review {
    id:                 number;
    node_id:            string;
    user:               User;
    body:               string;
    commit_id:          string;
    submitted_at:       Date;
    state:              string;
    html_url:           string;
    pull_request_url:   string;
    author_association: AuthorAssociation;
    _links:             ReviewLinks;
}

export interface ReviewLinks {
    html:         Comments;
    pull_request: Comments;
}

export interface IGithubEventRepo {
    id:   number;
    name: string;
    url:  string;
}